tuftool: Remove KeySource enum, use tough-ssm and KeySource trait
This is a decent set of changes to tuftool. It fully removes the
KeySource enum in source.rs and instead uses the KeySource trait
from tough.

This commit also removes all SSM related code in favor of the
tough-ssm crate (which is a copy of this code).

It also removes the deref.rs file, as it is no longer need because
these features have stabilized in upstream Rust. See:
https://github.com/rust-lang/rust/issues/50264
and
https://github.com/rust-lang/rust/pull/64708